requirements.txt
django
distutils
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
decorator==5.1.1
idna==3.7
pycparser==2.21
python-dotenv==1.0.1
python-telegram-bot==13.0
pytz==2023.3.post1
requests==2.32.3
requests-toolbelt==0.10.1
setuptools==69.0.2
six==1.16.0
tornado==6.4
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15